+++ CHANGELOG.md
@@ -6,7 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+### Changed
+- 
 ## [v1.10.3] - 2026-03-25
 ### Security
 - [#98] Update `google.golang.org/grpc` to fix CVE-2026-33186